renamed button file to .jsx, and fixed up button stories
