--- packages/design-system/src/stories/Button.stories.js
@@ -1,38 +1,0 @@
-import { BaseButton } from '@layer5/sistent-components';
-// More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
-export default {
-  title: 'Example/Button',
-  component: BaseButton,
-  tags: ['autodocs'],
-  argTypes: {
-    backgroundColor: { control: 'color' }
-  }
-};
-
-// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
-export const Primary = {
-  args: {
-    primary: true,
-    label: 'Button'
-  }
-};
-
-export const Secondary = {
-  args: {
-    label: 'Button'
-  }
-};
-
-export const Large = {
-  args: {
-    size: 'large',
-    label: 'Button'
-  }
-};
-
-export const Small = {
-  args: {
-    size: 'small',
-    label: 'Button'
-  }
-};
